requirements.txt
python-telegram-bot==22.5
python-dotenv==1.0.1
google-genai==1.1.0
chromadb>=0.4.22
langchain==0.3.13
langchain-google-genai==2.0.8
langchain-community==0.3.13
pypdf==5.1.0
pydub>=0.25.1
